[SPARK-13296][SQL] Move UserDefinedFunction into sql.expressions package.
